requirements.txt
numpy
librosa
opuslib==3.0.1
PyAudio==0.2.11
pycryptodome==3.12.0
websocket_client==1.2.3
